refactor: dialogs.py app_setting -> components/dialogs/
