+++ chronohelper/ui/components/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+"""
+UI組件
+"""
+
+# 導出SettingsDialog
+from chronohelper.ui.components.dialogs.settings_dialog import SettingsDialog
+
+__all__ = ['SettingsDialog']
