+++ chronohelper/ui/components/dialogs/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+"""
+對話框UI組件
+"""
+
+# 導出所有對話框類
+from chronohelper.ui.components.dialogs.settings_dialog import SettingsDialog
+
+__all__ = ['SettingsDialog']
